Converting Browse page and components to work with pagination and back-end filtering, moving /objectutils.js into /utils dir.
