+++ src/encoded/static/components/util/object.js
@@ -1,0 +1,62 @@
+'use strict';
+
+var _ = require('underscore');
+
+
+module.exports = {
+
+    /**
+     * Find property within an object using a propertyName in object dot notation.
+     * Recursively travels down object tree following dot-delimited property names.
+     * If any node is an array, will return array of results.
+     * 
+     * @param {Object} object - Item to traverse or find propertyName in.
+     * @param {string|string[]} propertyName - (Nested) property in object to retrieve, in dot notation or ordered array.
+     * @return {*} - Value corresponding to propertyName.
+     */
+
+    getNestedProperty : function(object, propertyName, suppressNotFoundError = false){
+
+        if (typeof propertyName === 'string') propertyName = propertyName.split('.'); 
+        if (!Array.isArray(propertyName)) throw new Error('Using improper propertyName in objectutils.getNestedProperty.');
+
+        return (function findNestedValue(currentNode, fieldHierarchyLevels, level = 0){
+            if (level == fieldHierarchyLevels.length) return currentNode;
+
+            if (Array.isArray(currentNode)){
+                var arrayVals = [];
+                for (var i = 0; i < currentNode.length; i++){
+                    arrayVals.push( findNestedValue(currentNode[i], fieldHierarchyLevels, level) );
+                }
+                return arrayVals;
+            } else {
+                if (typeof object === 'undefined' || !object) {
+                    if (!suppressNotFoundError) throw new Error('Field ' + _.clone(fieldHierarchyLevels).splice(0, level + 1).join('.') + ' not found on object.');
+                    return;
+                }
+                return findNestedValue(
+                    currentNode[fieldHierarchyLevels[level]],
+                    fieldHierarchyLevels,
+                    level + 1
+                );
+            }
+        })(object, propertyName);
+
+    },
+
+    singleTreatment : function(treatment) {
+        var treatmentText = '';
+
+        if (treatment.concentration) {
+            treatmentText += treatment.concentration + (treatment.concentration_units ? ' ' + treatment.concentration_units : '') + ' ';
+        }
+        treatmentText += treatment.treatment_term_name + (treatment.treatment_term_id ? ' (' + treatment.treatment_term_id + ')' : '') + ' ';
+        if (treatment.duration) {
+            treatmentText += 'for ' + treatment.duration + ' ' + (treatment.duration_units ? treatment.duration_units : '');
+        }
+        return treatmentText;
+    }
+
+};
+
+
