+++ src/encoded/static/libs/react-middleware.js
@@ -7,9 +7,8 @@ var fs = require('fs');
 var inline = fs.readFileSync(__dirname + '/../build/inline.js', 'utf8').toString();
 var store = require('../store');
 var { Provider, connect } = require('react-redux');
-var { JWT } = require('../components/objectutils');
+var { JWT, expFilters } = require('../components/util');
 var Alerts = require('../components/alerts');
-var hrefToFilters = require('../components/facetlist').hrefToFilters;
 
 var render = function (Component, body, res) {
     //var start = process.hrtime();
@@ -18,7 +17,7 @@ var render = function (Component, body, res) {
         'context': context,
         'href': res.getHeader('X-Request-URL') || context['@id'],
         'inline': inline,
-        'expSetFilters' : hrefToFilters(res.getHeader('X-Request-URL') || context['@id'])
+        'expSetFilters' : expFilters.hrefToFilters(res.getHeader('X-Request-URL') || context['@id'])
     };
 
     // Subprocess-middleware re-uses process on prod. Might have left-over data from prev request.
